+++ app/event/members.js
@@ -1,5 +1,5 @@
-const insertMembers = require("../db/members_insert.js");
-const getMember = require("../db/members_select.js");
+const insertMembers = require("../../db/members_insert.js");
+const getMember = require("../../db/members_select.js");
 
 module.exports = async function chatCountUp(msg) {
     let id = msg.author.id;
